+++ mockserver/src/main/java/com/byoutline/mockserver/HttpMockServer.java
@@ -1,9 +1,11 @@
 package com.byoutline.mockserver;
 
+import com.byoutline.mockserver.internal.AutoPortConnect;
 import com.byoutline.mockserver.internal.ConfigParser;
 import com.byoutline.mockserver.internal.MockNetworkLag;
 import com.byoutline.mockserver.internal.ParsedConfig;
 import com.byoutline.mockserver.internal.ResponseHandler;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.simpleframework.http.Request;
@@ -12,37 +14,34 @@ import org.simpleframework.http.core.Container;
 import org.simpleframework.http.core.ContainerServer;
 import org.simpleframework.transport.Server;
 import org.simpleframework.transport.connect.Connection;
-import org.simpleframework.transport.connect.SocketConnection;
+
+import java.io.IOException;
 
 import javax.annotation.Nonnull;
-import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.net.SocketAddress;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  * Local mock HTTP server.
  */
 public class HttpMockServer implements Container {
 
-    private final static Logger LOGGER = Logger.getLogger(HttpMockServer.class.getName());
-
     public static boolean DEBUG = true;
     private final Connection conn;
     private final ResponseHandler responseHandler;
 
+    // Reference is kept only to prevent server from being garbage collected
     private static HttpMockServer sMockServer;
 
-    public HttpMockServer(@Nonnull JSONObject jsonObject, @Nonnull ConfigReader configReader, @Nonnull NetworkType simulatedNetworkType)
-            throws IOException, JSONException {
+    public HttpMockServer(
+            @Nonnull JSONObject jsonObject,
+            @Nonnull ConfigReader configReader,
+            @Nonnull NetworkType simulatedNetworkType
+    ) throws IOException, JSONException {
         ParsedConfig config = new ConfigParser(configReader).parseConfig(jsonObject);
         MockNetworkLag networkLag = new MockNetworkLag(simulatedNetworkType);
         this.responseHandler = new ResponseHandler(config.responses, networkLag, configReader);
         Server server = new ContainerServer(this);
-        conn = new SocketConnection(server);
-        final SocketAddress sa = new InetSocketAddress(config.port);
-        conn.connect(sa);
+
+        conn = new AutoPortConnect().connectToPortFromRange(server, config.port, config.maxPort);
     }
 
     /**
@@ -51,18 +50,16 @@ public class HttpMockServer implements Container {
      * @param configReader         wrapper for platform specific bits
      * @param simulatedNetworkType delay time before response is sent.
      */
-    public static HttpMockServer startMockApiServer(@Nonnull ConfigReader configReader,
-                                                    @Nonnull NetworkType simulatedNetworkType) {
-        try {
-            JSONObject jsonObj = ConfigParser.getMainConfigJson(configReader);
-            sMockServer = new HttpMockServer(jsonObj, configReader, simulatedNetworkType);
-            return sMockServer;
-        } catch (IOException e) {
-            LOGGER.log(Level.SEVERE, "MockServer error:", e);
-        } catch (JSONException e) {
-            LOGGER.log(Level.SEVERE, "MockServer error:", e);
-        }
-        return null;
+    public static HttpMockServer startMockApiServer(
+            @Nonnull ConfigReader configReader,
+            @Nonnull NetworkType simulatedNetworkType
+    ) throws IOException {
+        JSONObject jsonObj = ConfigParser.getMainConfigJson(configReader);
+        HttpMockServer mockServer;
+        mockServer = new HttpMockServer(jsonObj, configReader, simulatedNetworkType);
+
+        sMockServer = mockServer;
+        return mockServer;
     }
 
     public void reset() {
@@ -71,6 +68,7 @@ public class HttpMockServer implements Container {
 
     public void shutdown() throws Exception {
         conn.close();
+        sMockServer = null;
         this.reset();
     }