+++ mockserver/src/main/java/com/byoutline/mockserver/internal/MatchingMethod.java
@@ -1,8 +1,5 @@
 package com.byoutline.mockserver.internal;
 
-/**
- * @author Sebastian Kacprzak <sebastian.kacprzak at byoutline.com>
- */
 public enum MatchingMethod {
     EXACT("EXACT"), CONTAINS("CONTAINS"), NOT_CONTAINS("NOT_CONTAINS"), ANY("ANY");
     public final String configValue;